refactor: remove dead paste-image code from server.ts, share readBinaryBody
- Remove IMAGE_MIME_EXT, readBinaryBody, /api/paste-image route, and
  pruneOldPastedImages call from server.ts — server.ts is dead code in
  --web mode (never spawned by tui/index.ts anymore).
- Move readBinaryBody into paste-utils.ts and export it; import it in
  frontend-daemon.ts instead of duplicating the definition.
- Fix byte-count comment in e2e test (68 bytes, not 67).
